login_mvc_project/requirements.txt
PySide6>=6.6

#
pyserial>=3.5